% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{EvaluationMeasures.table}
\alias{EvaluationMeasures.table}
\title{EvaluationMeasures.table}
\usage{
EvaluationMeasures.table(Real, Predicted, Positive = 1)
}
\arguments{
\item{Real}{Real binary values of the class}

\item{Predicted}{Predicted binary values of the class}

\item{Positive}{Consider 1 label as Positive Class unless changing this parameter to 0}
}
\value{
TP,TN,FP,FN
}
\description{
Specify the number of TP,TN,FP,FN
}
\details{
By getting the predicted values and real values calulate the number of True positive samples, False Negative, False Positive and True Negative
}
\examples{
EvaluationMeasures.table(c(1,0,1,0,1,0,1,0),c(1,1,1,1,1,0,0,0))
}
\author{
Babak Khorsand
}

